+++ src/client/MainWindow.h
@@ -36,7 +36,7 @@ private:
     void buildDefaultLoginPage();
     void ensureSystemPages();
     void wirePageKeypad(const QString &pageName);
-    void handleAction(const QString &pageName, ActionType action);
+    void handleAction(const QString &pageName, ActionType action, const QString &targetPage);
     void toggleEditMode();
     void openPropertyDialog(UiElement *elem);
     void openPageManager();